Move API calls to package
